+++ linux-notebook-installer.sh
@@ -30,3 +30,7 @@ eval $(/home/linuxbrew/.linuxbrew/bin/brew shellenv)
 
 heading "Installing Homebrew tools ..."
 brew install gh
+
+heading "Installing Signal ..."
+sudo snap install signal-desktop
+