+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # The Changelog
 
+## [0.0.6] (January 30 2021)
+
+**Updated :**
+- Rename project to: The Linux Installer
+
+
 ## [0.0.5] (December 28 2020)
 
 **Added :**
@@ -16,11 +22,13 @@
 
 - from: #!/bin/sh to: #!/bin/bash
 
+
 ## [0.0.4] (December 27 2020)
 
 **Added :**
 
 - Homebrew to PATH
+
 
 ## [0.0.3] (December 24 2020)
 
@@ -33,11 +41,13 @@
 
 - script commands syntax
 
+
 ## [0.0.2] (December 23 2020)
 
 **Added :**
 
 - Homebrew package manager
+
 
 ## [0.0.1] (December 22 2020)